+++ projects/data/src/lib/stat-aggregators/clan-roster.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { ClanRosterService } from './clan-roster.service';
+
+describe('ClanRosterService', () => {
+  let service: ClanRosterService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(ClanRosterService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});